enable CLANG_WARN_SUSPICIOUS_IMPLICIT_CONVERSION
And fix the problems that become visible:

Mostly by using CGFloat (double) instead of float.

In a few cases by making the cast to int explicit,
not using integers as booleans, booleans as
pointers and using the precisely typed [l]ceil[f]
functions required.

+++ Classes/Controllers/DBPrefsWindowController.m
@@ -376,7 +376,7 @@ static DBPrefsWindowController *_sharedPrefsWindowController = nil;
 {
 	NSRect windowFrame = [[self window] frame];
 	NSRect contentRect = [[self window] contentRectForFrameRect:windowFrame];
-	float windowTitleAndToolbarHeight = NSHeight(windowFrame) - NSHeight(contentRect);
+	CGFloat windowTitleAndToolbarHeight = NSHeight(windowFrame) - NSHeight(contentRect);
 
 	windowFrame.size.height = NSHeight([view frame]) + windowTitleAndToolbarHeight;
 	windowFrame.size.width = NSWidth([view frame]);
