+++ Classes/Views/PBGitRevisionCell.h
@@ -20,7 +20,7 @@
 	__weak IBOutlet id<PBRefContextDelegate> contextMenuDelegate;
 }
 
-- (int) indexAtX:(float)x;
+- (int) indexAtX:(CGFloat)x;
 - (NSRect) rectAtIndex:(int)index;
 - (void) drawLabelAtIndex:(int)index inRect:(NSRect)rect;
